attempt 2

+++ c/main_project/led_control/src/server.c
@@ -32,7 +32,7 @@ const char *get_content_type(const char *path) {
 int request_handler(void *cls, struct MHD_Connection *connection,
     const char *url, const char *method,
     const char *version, const char *upload_data,
-    size_t *upload_data_size, void **con_cls) {
+    long unsigned int *upload_data_size, void **con_cls) {
 
   if (strcmp(method, "GET") == 0) {
   return handle_get_request(connection, url);
